testing out Context Engine MCP
got an email:
Now live in beta!
Hi there,
We noticed recently you've switched to a different coding agent, but we've been working on something exciting that we wanted to share with you!
Our Context Engine MCP is available in beta for existing customers, we would love for you to be one of the first to try it out using the agent of your choice.
 Try this out today 

We've seen positive results so far in our early testing with GitHub copilot, Kilo code and Opencode.
As part of our beta launch, customers who try our new MCP will have free usage to experiment with it in their workflow. We only ask that you provide feedback via a short survey at the end of the week.